Rename full VIC assembly implementation header
Rename vic_solve_full_impl.h to vic_assemble_full_impl.h so the header name matches the vic_assemble_full_impl function it provides.

Update CPU and CUDA implicit dispatch includes to use the new header name. No behavior changes.

Validation: not rerun for this rename-only commit; full validation is run after the follow-up dispatch refactor commit.
