+++ src/implicit/implicit_dispatch.cpp
@@ -13,9 +13,9 @@
 // snap
 #include "forward_sweep_impl.h"
 #include "implicit_dispatch.hpp"
+#include "vic_assemble_full_impl.h"
 #include "vic_assemble_partial_impl.h"
 #include "vic_redistribute_impl.h"
-#include "vic_solve_full_impl.h"
 
 namespace snap {
