+++ src/implicit/implicit_dispatch.cu
@@ -16,7 +16,7 @@
 #include "implicit_dispatch.hpp"
 #include "vic_assemble_partial_impl.h"
 #include "vic_redistribute_impl.h"
-#include "vic_solve_full_impl.h"
+#include "vic_assemble_full_impl.h"
 
 namespace snap {
