+++ src/implicit/vic_assemble_full_impl.h
